+++ common/src/main/java/com/faboslav/featurify/common/api/FeaturifyPlacedFeature.java
@@ -9,5 +9,4 @@ public interface FeaturifyPlacedFeature
 
 	@Nullable
 	Identifier featurify$getIdentifier();
-
 }