+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.0.2
+
+- Added configuration of weighted (sub)features
+- Fixed load/save
+- Removed more debug messages
+
 ## 1.0.1
 
 - Fixed crash related to other mods using custom generation steps